% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{Arrange_Table}
\alias{Arrange_Table}
\title{Arrange Table Function - Sort and filter columns of a data frame or matrix by a specified vector}
\usage{
Arrange_Table(Mat, FilterList)
}
\arguments{
\item{Mat}{Target matrix or df.}

\item{FilterList}{Accord which to sort and filter columns}
}
\value{
a Matrix
}
\description{
Arrange Table Function - Sort and filter columns of a data frame or matrix by a specified vector
}
\author{
Zhiming Ye
}
