% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{CalcMad}
\alias{CalcMad}
\title{Output expression matrix based on top MAD value}
\usage{
CalcMad(mat, num)
}
\arguments{
\item{mat}{Expression matrix}

\item{num}{select how many genes}
}
\value{
A matrix
}
\description{
Output expression matrix based on top MAD value
}
