% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{CorrlatinEstimate0}
\alias{CorrlatinEstimate0}
\title{Estimate correlation by batch}
\usage{
CorrlatinEstimate0(
  genegsvamatrix,
  TargetSet = "HALLMARK_HYPOXIA",
  UseMethod = "spearman"
)
}
\arguments{
\item{genegsvamatrix}{Target matrix or data frame.}

\item{TargetSet}{Target estimated column name}

\item{UseMethod}{default as "spearman"}
}
\value{
R and correlation coeffection
}
\description{
Estimate correlation by batch
}
