% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{countToFPKM}
\alias{countToFPKM}
\title{Convert counts to TPM}
\usage{
countToFPKM(Counts, Length, gene)
}
\arguments{
\item{Counts}{Expression matrix}

\item{Length}{length vector of gene}

\item{gene}{names of genes}
}
\value{
FPKM matrix
}
\description{
Convert counts to TPM
}
\author{
Zhiming Ye
}
