% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{do.GSVA}
\alias{do.GSVA}
\title{Do GSVA analysis}
\usage{
\method{do}{GSVA}(Mat, GMTfilePath)
}
\arguments{
\item{Mat}{Expression Matrix}

\item{GMTfilePath}{gmt file path from MsigDB and so on.}
}
\value{
A GSVA score matrix
}
\description{
Do GSVA analysis
}
