% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{do.ssGSEA}
\alias{do.ssGSEA}
\title{Do ssGSEA analysis}
\usage{
\method{do}{ssGSEA}(Mat, GMTfilePath, numOfCore = 2)
}
\arguments{
\item{Mat}{Expression Matrix}

\item{GMTfilePath}{gmt file path from MsigDB and so on.}

\item{numOfCore}{use how many CPU cores to calculate.}
}
\value{
A enrichment score matrix
}
\description{
Do ssGSEA analysis
}
