% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{doGO}
\alias{doGO}
\title{ORA GO analysis by clusterProfiler by symbol identifiers.}
\usage{
doGO(GS, ont = "BP", PVal = 0.01, QVal = 0.05)
}
\arguments{
\item{GS}{Gene Symbols for enrichment analysis. NOT ENSEMBL ID.}

\item{ont}{One of "BP", "CC", "MF"  and "ALL".}

\item{PVal}{P adjusted value cut off, default 0.01}

\item{QVal}{qvalue cut off, default 0.05}
}
\value{
An enrichment result generated by clusterProfiler.
}
\description{
ORA GO analysis by clusterProfiler by symbol identifiers.
}
