% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{doKEGG}
\alias{doKEGG}
\title{ORA KEGG analysis by clusterProfiler by symbol identifiers.}
\usage{
doKEGG(GS, PVal = 0.01, QVal = 0.05, OrgDB = org.Hs.eg.db)
}
\arguments{
\item{GS}{Gene Symbols for enrichment analysis. NOT ENSEMBL ID.}

\item{PVal}{P adjusted value cut off, default 0.01}

\item{QVal}{qvalue cut off, default 0.05}

\item{OrgDB}{default as org.Hs.eg.db}
}
\value{
An enrichment result generated by clusterProfiler.
}
\description{
ORA KEGG analysis by clusterProfiler by symbol identifiers.
}
\author{
Zhiming Ye
}
