% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{doRA}
\alias{doRA}
\title{ORA Reactome analysis by clusterProfiler by symbol identifiers.}
\usage{
doRA(GS, PVal = 0.01, QVal = 0.05)
}
\arguments{
\item{GS}{Gene Symbols for enrichment analysis. NOT ENSEMBL ID.}

\item{PVal}{P adjusted value cut off, default 0.01}

\item{QVal}{qvalue cut off, default 0.05}
}
\value{
An enrichment result generated by clusterProfiler.
}
\description{
ORA Reactome analysis by clusterProfiler by symbol identifiers.
}
