% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{MapSymbol}
\alias{MapSymbol}
\title{Convert Ensembl ID to Symbol}
\usage{
MapSymbol(Mat, ColNum)
}
\arguments{
\item{Mat}{Target expression matrix}

\item{ColNum}{Number of column contains ensembl ID}
}
\value{
a matrix
}
\description{
Convert Ensembl ID to Symbol
}
\author{
Zhiming Ye
}
