% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{MultiDunn}
\alias{MultiDunn}
\title{Multi group nonparametric test}
\usage{
MultiDunn(Mat, FactorCol, Method = "bonferroni", OnlySig = F, pcutoff = 0.01)
}
\arguments{
\item{Mat}{Columns should be every variable to be tested, and another one column contains information about grouping, which is refer to the FactorCol}

\item{FactorCol}{Which column contains the grouping information. The number of it.}

\item{Method}{Default the "bonferroni"}

\item{OnlySig}{Whether show only significant result.}

\item{pcutoff}{P value cut off.}
}
\description{
Multi group nonparametric test
}
\author{
Zhiming Ye
}
