% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{NMFEstimate}
\alias{NMFEstimate}
\title{NMF Estimation}
\usage{
NMFEstimate(mat, pdfname, pngname, filename)
}
\arguments{
\item{mat}{Matrix}

\item{pdfname}{Output PDF location}

\item{pngname}{Output PNG location}

\item{filename}{Output RData location}
}
\description{
NMF Estimation
}
