% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{NormalizeTest}
\alias{NormalizeTest}
\title{Perform a normality test by batch}
\usage{
NormalizeTest(Table)
}
\arguments{
\item{Table}{Target table}
}
\value{
P Val
}
\description{
Perform a normality test by batch
}
\author{
Zhiming Ye
}
