% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{Plot.Box.Lite}
\alias{Plot.Box.Lite}
\title{Two-group Box plot with stastical test}
\usage{
Plot.Box.Lite(
  Mat,
  xcolName,
  ycolName,
  colorList,
  my_comparisons,
  Method = "wilcox.test"
)
}
\arguments{
\item{Mat}{A table}

\item{xcolName}{Name of group 1}

\item{ycolName}{Name of group 2}

\item{colorList}{A vector of colors}

\item{my_comparisons}{Comparison list}

\item{Method}{like "wilcox.test" and "t.test"}
}
\value{
A ggplot2 object
}
\description{
Two-group Box plot with stastical test
}
\author{
Zhiming Ye
}
