% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{Plot.lm.smooth}
\alias{Plot.lm.smooth}
\title{Plot the fitted line}
\usage{
Plot.lm.smooth(dataset, xval, yval, Method = "spearman")
}
\arguments{
\item{dataset}{variables should be columns.}

\item{Method}{like "pearson" or "spearman"}
}
\value{
A ggplot2 object
}
\description{
Plot the fitted line
}
