% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{PlotNES.lite}
\alias{PlotNES.lite}
\title{Generate a GSEA plot with ranked curve, which only used when drawing multi-pathways.}
\usage{
PlotNES.lite(gseRes, GeneSet, nCols = 2, AnnoTerms = c("NES", "p.adjust"))
}
\arguments{
\item{gseRes}{GSEA result generated by clusterProfiler}

\item{GeneSet}{Description of the selected pathway. A vector.}

\item{nCols}{layout of figure, numbers of columns.}

\item{AnnoTerms}{Default the NES and p.adjust}
}
\value{
A cowplot plot_gird object
}
\description{
Generate a GSEA plot with ranked curve, which only used when drawing multi-pathways.
}
\author{
Zhiming Ye
}
