% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{PlotPCA}
\alias{PlotPCA}
\title{A simple PCA plot generator.}
\usage{
PlotPCA(Mat, group, Ellipse = T, counts = F, loged = T, NumOfGenes = 10000)
}
\arguments{
\item{Mat}{Expression levels.}

\item{group}{A vector, The order shall be consistent with the table.}

\item{counts}{If the expression table is original counts, should be TRUE. If is the TPM, FPKM or RPKM, should be FALSE.}

\item{loged}{Whether the expression level is log-transformed.}

\item{NumOfGenes}{Choose the top n genes with the highest MAD value for PCA analysis.}
}
\value{
A ggplot2 object
}
\description{
A simple PCA plot generator.
}
