% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{PlotPie}
\alias{PlotPie}
\title{Generate pie plot}
\usage{
PlotPie(
  dt,
  itemAndfreq = c(1, 2),
  pretty = T,
  Tag = "B",
  Order = "no",
  col_type = "nejm"
)
}
\arguments{
\item{dt}{A data frame contain items and frequency}

\item{itemAndfreq}{Vector specify the items and frequency, should be numeric.}

\item{Tag}{The label}

\item{Order}{Order, a character vector}

\item{col_type}{npg or lancet or nejm}
}
\value{
A ggplot2 object
}
\description{
Generate pie plot
}
\author{
Zhiming Ye
}
