% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{PlotSanky}
\alias{PlotSanky}
\title{Generate Sankey Diagram.}
\usage{
PlotSanky(InfoTable, ItemA, ItemB, ItemAlist, ItemBlist)
}
\arguments{
\item{InfoTable}{A table with specific 2 columns which brings Categorical variable.}

\item{ItemA}{Column name of item A.}

\item{ItemB}{Column name of item B.}

\item{ItemAlist}{specified ordering, a character vector.}

\item{ItemBlist}{specified ordering, a character vector.}
}
\value{
A ggplot2 object
}
\description{
Generate Sankey Diagram.
}
