% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{PrettyMsigdbName}
\alias{PrettyMsigdbName}
\title{Remove "_" in MsigDB items}
\usage{
PrettyMsigdbName(CHAR, start)
}
\arguments{
\item{CHAR}{vector of MsigDB item names}

\item{start}{number of character which is used to start to separate.}
}
\value{

}
\description{
Remove "_" in MsigDB items
}
