% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{Ranked.GS}
\alias{Ranked.GS}
\title{Return ranked gene list which is use for "GSEA" and "cnetplot" in package "clusterProfiler"}
\usage{
Ranked.GS(Gene, log2FC, FromType = "SYMBOL")
}
\arguments{
\item{FromType}{default is "SYMBOL"}
}
\value{
A ranked named numeric vector. Names of the numbers is the ENTREZID.
}
\description{
Return ranked gene list which is use for "GSEA" and "cnetplot" in package "clusterProfiler"
}
