% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{rmEnsemblDot}
\alias{rmEnsemblDot}
\title{remove the Ensembl ID}
\usage{
rmEnsemblDot(Mat, ColNum)
}
\arguments{
\item{Mat}{Expression Matrix from Xena and so on}

\item{ColNum}{Which column contains ensembl ID, as numeric.}
}
\value{
A data frame
}
\description{
remove the Ensembl ID
}
