% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{scanpy.QueryDEG}
\alias{scanpy.QueryDEG}
\title{Extract Marker genes of scanpy DEG Result}
\usage{
scanpy.QueryDEG(
  DEGmat,
  QueryGene,
  numOfClusters,
  verbose = T,
  PvalCutoff = 0.01,
  lFCcutoff = 1,
  IsHighExpGenes = T,
  geneColName = "_names",
  pvalColName = "_pvals_adj",
  lfcColName = "_logfoldchanges"
)
}
\arguments{
\item{DEGmat}{Result of DEG}

\item{QueryGene}{Gene Symbol}

\item{numOfClusters}{Number of cell clusters}

\item{verbose}{Print Pval and log2FC}

\item{PvalCutoff}{Pval cut off}

\item{lFCcutoff}{log2FC cut off}

\item{IsHighExpGenes}{Choose whether FC>0 or <0}

\item{lfcColName}{}
}
\value{
Print the chosen cell cluster
}
\description{
The result is generated by pandas using for command. Ref to the scanpy document.

`pd.DataFrame(`




}
\author{
Zhiming Ye
}
