% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{UniCox}
\alias{UniCox}
\title{Batch uni-Cox regression to find survival-related genes}
\usage{
UniCox(StatusAndTimeCol, DATA)
}
\arguments{
\item{StatusAndTimeCol}{Numeric vector, item first is number of status column and the second is the time column}

\item{DATA}{Data set}
}
\value{
A Cox-regression result
}
\description{
Batch uni-Cox regression to find survival-related genes
}
\author{
Zhiming Ye
}
