% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{WGCNACalcSFT}
\alias{WGCNACalcSFT}
\title{Calculate the SFT value for WGCNA analysis}
\usage{
WGCNACalcSFT(Mat, OutPutFile)
}
\arguments{
\item{Mat}{Expression matrix}

\item{OutPutFile}{Output figure location}
}
\description{
Calculate the SFT value for WGCNA analysis
}
