% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{wilcox.adj.test}
\alias{wilcox.adj.test}
\title{Batch wilCox test, with P value Corrected.}
\usage{
wilcox.adj.test(Mat, FactorCol, Method = "BH")
}
\arguments{
\item{Mat}{Columns should be every variable to be tested, and another one column contains information about grouping, which is refer to the FactorCol}

\item{FactorCol}{Which column contains the grouping information. The number of it.}

\item{Method}{Default is "BH".}
}
\description{
Batch wilCox test, with P value Corrected.
}
\author{
Zhiming Ye
}
