% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{Write.tables}
\alias{Write.tables}
\title{Write tab table file}
\usage{
Write.tables(table, withRowname = T)
}
\arguments{
\item{table}{A character, name of data frame or matrix object.}

\item{withRowname}{with row names or not.}
}
\value{
NULL, save a file to work dictionary.
}
\description{
Write tab table file
}
\author{
Zhiming Ye
}
