% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{MapSymbol}
\alias{MapSymbol}
\title{Convert Ensembl ID to Symbol}
\usage{
MapSymbol(TargetGene, ColNum = 1, OrgDB = org.Hs.eg.db, Type = "to")
}
\arguments{
\item{TargetGene}{Target gene ID}

\item{ColNum}{Number of column contains ensembl ID, only nesessary when input a matrix}

\item{OrgDB}{The annotation DB, default as org.Hs.eg.db}

\item{Type}{to:ensembl ID->symbol, from:symbol->ID}
}
\value{
a matrix
}
\description{
Convert Ensembl ID to Symbol
}
\author{
Zhiming Ye
}
