% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{PlotNES}
\alias{PlotNES}
\title{A useful plot with NES and pvalue to display the result of GSEA analysis}
\usage{
PlotNES(
  gseRes,
  showItemsNum,
  Col1 = "#B39CD0",
  Col2 = "#FBEAFF",
  Ratio = 0.5,
  SepCol = F,
  HiExprGroupName = "A",
  LoExprGroupName = "B",
  WarpWidth = 50
)
}
\arguments{
\item{gseRes}{GSEA result generated by clusterProfiler}

\item{showItemsNum}{Numbers of shown items.}

\item{Col1}{Color 1}

\item{Col2}{Color 2}

\item{Ratio}{Ration of numbers of up regulated pathways to down regulated pathways.}

\item{SepCol}{Method of display result. Display according groups or P values. if True display by groups, false displayed by pval.default is false.}

\item{HiExprGroupName}{Char vector}

\item{LoExprGroupName}{Char vector}

\item{WarpWidth}{Wrap length}
}
\value{
a ggplot2 object.
}
\description{
A useful plot with NES and pvalue to display the result of GSEA analysis
}
\author{
Zhiming Ye
}
