% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{rmSep}
\alias{rmSep}
\title{remove separate from character}
\usage{
rmSep(Chr, SepChr)
}
\arguments{
\item{Chr}{Target character vector}

\item{SepChr}{seperate character}
}
\value{
a character vector
}
\description{
remove separate from character
}
